fix: view 분리
